pyproject.toml
[project]
name = "referral-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "easyocr>=1.7.2",
    "numpy>=2.3.2",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "torch>=2.8.0",
    "transformers>=4.55.2",
]
